requirements.txt
pyserial==3.4
netifaces==0.10.4

pytest==3.2.2
grpcio==1.6.3
grpcio-tools==1.6.3
